Update main.py
Created functions for choices made by the user as well as taking screenshots to find a pixel color. 
Added a second option for when steam asks for mobile verification. It goes through and puts every item on the market and then goes to the next 25 items.
Resolution function was made on a whim, I have not tested it, however from what I understand it will NOT work since I move the mouse around by exact pixel coordinates and they would be different per resolution.